Removes Support for REST in favor of SOAP
