--- lib/zuora/models/charge_update.rb
@@ -1,13 +1,0 @@
-class Zuora::Models::ChargeUpdate
-  include SchemaModel
-
-  schema :charge_update,
-    rate_plan_charge_id: {
-      doc: 'Subscription ID',
-      type: String
-    },
-
-    quantity: {
-      type: Numeric
-    }
-end
