+++ lib/zuora/object.rb
@@ -1,0 +1,5 @@
+module Zuora
+  class Object < Hash
+    include Hashie::Extensions::MethodAccess
+  end
+end
