--- lib/zuora/soap/object.rb
@@ -1,7 +1,0 @@
-module Zuora
-  module Soap
-    class Object < Hash
-      include Hashie::Extensions::MethodAccess
-    end
-  end
-end
