--- spec/cassettes/cancel_subscription_failure.yml
@@ -1,110 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://apisandbox-api.zuora.com/rest/v1/connections
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Apiaccesskeyid:
-      - engineering+dev@contactually.com
-      Apisecretaccesskey:
-      - "<ZUORA_SANDBOX_PASSWORD>"
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '0'
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '22'
-      Expires:
-      - Thu, 21 Jan 2016 02:01:59 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Thu, 21 Jan 2016 02:01:59 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=5RVhWdL3okMtJBeinFMRnhlWlAqF_9DDqWviy5wvTq39k4Is15Rr0pB_LGjlzv7nik0u_RANRgHqA8wpSqFBNF-_yE5lBAL1sdb3TIX7eK5mY8JEyDDynt7ZsOR6s-n-3ebzt8zKB6ldpJ6Uwt9Raju4iRKYp_11A9KzsQWOzjgGboY5IuZmsFMhQfgn61PW4rWKBCKv_3954ZmpxKDksnwwBdU-MX5INDuXHsXDfQAedi1q-9hsJ7MgqWW2Jzj4;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : true
-        }
-    http_version: 
-  recorded_at: Thu, 21 Jan 2016 02:01:58 GMT
-- request:
-    method: put
-    uri: https://apisandbox-api.zuora.com/rest/v1/subscriptions/INVALID_ID/cancel
-    body:
-      encoding: UTF-8
-      string: '{"cancellationPolicy":"SpecificDate","cancellationEffectiveDate":"2016-01-20","invoiceCollect":false}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Content-Type:
-      - application/json
-      Cookie:
-      - ZSession=5RVhWdL3okMtJBeinFMRnhlWlAqF_9DDqWviy5wvTq39k4Is15Rr0pB_LGjlzv7nik0u_RANRgHqA8wpSqFBNF-_yE5lBAL1sdb3TIX7eK5mY8JEyDDynt7ZsOR6s-n-3ebzt8zKB6ldpJ6Uwt9Raju4iRKYp_11A9KzsQWOzjgGboY5IuZmsFMhQfgn61PW4rWKBCKv_3954ZmpxKDksnwwBdU-MX5INDuXHsXDfQAedi1q-9hsJ7MgqWW2Jzj4;
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '166'
-      Expires:
-      - Thu, 21 Jan 2016 02:01:59 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Thu, 21 Jan 2016 02:01:59 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=yawvURxKihsIUshDbmWzbtZI6KAc1ZVQckYJnibIkGSqKnEjTg0aBVObqNfvyz0czlImEQT8wvWign7yYcTbZdfovzZiFvQq6dVDtt1oii8OMGM_N6wMnT3Kfmrhlae4GvIYufc-fJSJTTpNqHZSC1Tq2FNqclFUewDIvmXKv9AsMHe6ZFpsksNzeFu34lO1ijLORwuoofBWhnS_3sC3GycXwtvktkCvHFv9d-wO3ao1htpeov8eUIECMf8ZZH6Z;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : false,
-          "processId" : "0374B3C89E3CFE71",
-          "reasons" : [ {
-            "code" : 53200040,
-            "message" : "Cannot find entity by key: 'INVALID_ID'."
-          } ]
-        }
-    http_version: 
-  recorded_at: Thu, 21 Jan 2016 02:01:59 GMT
-recorded_with: VCR 3.0.1
