--- spec/cassettes/create_account_failure.yml
@@ -1,113 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://apisandbox-api.zuora.com/rest/v1/connections
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Apiaccesskeyid:
-      - engineering+dev@contactually.com
-      Apisecretaccesskey:
-      - "<ZUORA_SANDBOX_PASSWORD>"
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '0'
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '22'
-      Expires:
-      - Wed, 20 Jan 2016 22:09:27 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:09:27 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=6TgJ61nYFRw5QWHJuI5F-uHW0CLKQhe0UnVjFqe4PccIMN_KVgzWTEn3Orw30fX9EVIx6vZjJ2RZjtExQswBMVPNtLDNMY0UWr_Ia-sYxVMlBhHWGjl_Fta0uVDYzPtjeEHc69PJDeonPca0wc-qnmYmoqAANvI9rXReU3osSXlPoiE3cgpv1_hkB8oyrp_G6ZJ33H0XUCLWb8Ln19Qr0D1Y9xzK2TTImEb7avrxSWPPYRjU0P-z3VR3WwFDsfOC;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : true
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:09:26 GMT
-- request:
-    method: post
-    uri: https://apisandbox-api.zuora.com/rest/v1/accounts
-    body:
-      encoding: UTF-8
-      string: '{"name":"693d0a7a-7946-467a-90a6-cc9c7953588a","autoPay":true,"currency":"USD","billCycleDay":"0","paymentTerm":"Net
-        30","accountNumber":"123456","billToContact":{"firstName":"Velma","lastName":"Gutkowski","address1":"30661
-        Lynch Route","city":"Gillianshire","state":"NE","zipCode":"68108-3953","country":"USA"},"soldToContact":{"firstName":"Velma","lastName":"Gutkowski","address1":"30661
-        Lynch Route","city":"Gillianshire","state":"NE","zipCode":"68108-3953","country":"USA"},"creditCard":{"cardType":"Visa","cardNumber":"4242424242424242","expirationMonth":"03","expirationYear":2017,"securityCode":"122"}}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Content-Type:
-      - application/json
-      Cookie:
-      - ZSession=6TgJ61nYFRw5QWHJuI5F-uHW0CLKQhe0UnVjFqe4PccIMN_KVgzWTEn3Orw30fX9EVIx6vZjJ2RZjtExQswBMVPNtLDNMY0UWr_Ia-sYxVMlBhHWGjl_Fta0uVDYzPtjeEHc69PJDeonPca0wc-qnmYmoqAANvI9rXReU3osSXlPoiE3cgpv1_hkB8oyrp_G6ZJ33H0XUCLWb8Ln19Qr0D1Y9xzK2TTImEb7avrxSWPPYRjU0P-z3VR3WwFDsfOC;
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '163'
-      Expires:
-      - Wed, 20 Jan 2016 22:09:27 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:09:27 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=tQWpUiqi1Vhytd27GQv1Mg-K9rix9sf44CKVkrfARRU36R0gW_FGS2PC5jgKfu81xmRfO0V2JdGPx00prsf_529PvAtn-bA4U6VMpnIcSbasrxT7L-WcGSjzRA35OYenI3d4fj7_px-MJ1JNBhj5a8IfXtmMFUCLKm0pejKusVgEDHIRVqmxSkGMfCMXPFQc4ElgV-j6za2HeHGZkWcAlbzGPNcTKy2kRE1qZDngM1iN2lEEB-nG5_qdszMDtid7;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : false,
-          "processId" : "4798D6CB4997DF21",
-          "reasons" : [ {
-            "code" : 51000020,
-            "message" : "The account number 123456 is invalid."
-          } ]
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:09:27 GMT
-recorded_with: VCR 3.0.1
