--- spec/cassettes/create_subscription_failure.yml
@@ -1,111 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://apisandbox-api.zuora.com/rest/v1/connections
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Apiaccesskeyid:
-      - engineering+dev@contactually.com
-      Apisecretaccesskey:
-      - "<ZUORA_SANDBOX_PASSWORD>"
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '0'
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '22'
-      Expires:
-      - Wed, 20 Jan 2016 22:20:29 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:20:29 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=C20XpP_VIYSoErzQWSjD9nCOVlVcYaIKNTrCPfeCH-bH5uzzAqT9n9ZICdkkDF61RUO0yt1JFlZxs7_hJqc8oNOH6f_XPcuEHDcAWnRqzX9xWNB5BrhbEUjzWfUrwElQoJc6DCNLupmKp6ZKLkJ7jEAwjpCzR2UBQLe3hmSRKtbTRPkcDMofqjKBamp3ntDkyhwM-rPEok7Y-CEBy55M17cUHwbCyNxdpoRdebxSqeT57htqdswXRNW6qfO_ZZhQ;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : true
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:20:28 GMT
-- request:
-    method: post
-    uri: https://apisandbox-api.zuora.com/rest/v1/subscriptions
-    body:
-      encoding: UTF-8
-      string: '{"renewalTerm":3,"renewalTermPeriodType":"Week","termType":"EVERGREEN","invoiceTargetDate":"-4712-01-01","accountKey":"2c92c0fa525e25d50152611a674074a2","contractEffectiveDate":"2016-01-20","invoice":true,"collect":true,"initialTerm":12,"initialTermPeriodType":"Week","notes":"Test
-        POST subscription from zuora-ruby","serviceActivationDate":"2016-01-20","autoRenew":true,"subscribeToRatePlans":[{"productRatePlanId":"INVALID_ID"}]}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Content-Type:
-      - application/json
-      Cookie:
-      - ZSession=C20XpP_VIYSoErzQWSjD9nCOVlVcYaIKNTrCPfeCH-bH5uzzAqT9n9ZICdkkDF61RUO0yt1JFlZxs7_hJqc8oNOH6f_XPcuEHDcAWnRqzX9xWNB5BrhbEUjzWfUrwElQoJc6DCNLupmKp6ZKLkJ7jEAwjpCzR2UBQLe3hmSRKtbTRPkcDMofqjKBamp3ntDkyhwM-rPEok7Y-CEBy55M17cUHwbCyNxdpoRdebxSqeT57htqdswXRNW6qfO_ZZhQ;
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '166'
-      Expires:
-      - Wed, 20 Jan 2016 22:20:29 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:20:29 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=56FErn73qIjN-xTlUBvYFl9Ke8KFrNRqlFCYS3EJ61VLtGUzVYTxlegzFybEzc-e47KpN95jui6g0bPbrcye-K3-aQvUd5Viw2IzRgmJ29tBzSPfRE__RlT7sOrpqLTyUWp_9QKpLvxHO4RXyBiNwjaTIJM-7FybVuqcS9c-eX-LberWCKPYDugZgukPbf7LYzNQIdKorjPGwJpIpUSizs5UZh-JWVWdc95OPVOyb42yioJ0uSNxVFYB68y63otR;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : false,
-          "processId" : "6378106EC164D790",
-          "reasons" : [ {
-            "code" : 53000021,
-            "message" : "Invalid parameter(s): 'invoice,collect'."
-          } ]
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:20:29 GMT
-recorded_with: VCR 3.0.1
