--- spec/cassettes/create_subscription_success.yml
@@ -1,112 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://apisandbox-api.zuora.com/rest/v1/connections
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Apiaccesskeyid:
-      - engineering+dev@contactually.com
-      Apisecretaccesskey:
-      - "<ZUORA_SANDBOX_PASSWORD>"
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '0'
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '22'
-      Expires:
-      - Wed, 20 Jan 2016 22:21:35 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:21:35 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=_aiOm5PZDplkU7bnUJBGXQCvDhn395NXG-mtQ_xPC6D--4v-D6UrQvXoSHlf0-rNI7KEfapGpECM-s1Mc-dFxfsr4sGy0OwsTxuUqr-1d_vJ2S6lgDvFqGS2bCJ8YSBNZe9W5eMQO9ehhX_38s2S0eFstbKr14YWZN33hGYFdN_EGnJQh57BDmfCGnArvVrDqDfism8InNbnbsB3i_ha1op5MbcZQ8sLNYBHsjdlL5l-h-etjo9dVgWJwYi_0T_3;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : true
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:21:34 GMT
-- request:
-    method: post
-    uri: https://apisandbox-api.zuora.com/rest/v1/subscriptions
-    body:
-      encoding: UTF-8
-      string: '{"accountKey":"2c92c0fa525e25d50152611a674074a2","termType":"EVERGREEN","contractEffectiveDate":"2016-01-20","serviceActivationDate":"2016-01-20","autoRenew":true,"subscribeToRatePlans":[{"productRatePlanId":"2c92c0f950fa763f01510cbb937812dd"}]}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Content-Type:
-      - application/json
-      Cookie:
-      - ZSession=_aiOm5PZDplkU7bnUJBGXQCvDhn395NXG-mtQ_xPC6D--4v-D6UrQvXoSHlf0-rNI7KEfapGpECM-s1Mc-dFxfsr4sGy0OwsTxuUqr-1d_vJ2S6lgDvFqGS2bCJ8YSBNZe9W5eMQO9ehhX_38s2S0eFstbKr14YWZN33hGYFdN_EGnJQh57BDmfCGnArvVrDqDfism8InNbnbsB3i_ha1op5MbcZQ8sLNYBHsjdlL5l-h-etjo9dVgWJwYi_0T_3;
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '314'
-      Expires:
-      - Wed, 20 Jan 2016 22:21:36 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:21:36 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=zXgTN6euxLwvFAHdU1Cp9ROyZQzw5piItA83KCcCAuCfNNDpdVOoRWug1LU3UkQp5kc4TEWDrwtCuDInn_xjgDf4BBVNmJmYLTqkMVQiP51bz0GVVUacAZExg12rayKMXvcpPr8HGfVuSndR9V2iVMtVyWsGrTjoe4HYuzahYxd2SPo-ztHje4_WFaE3pphMhgzfPf_mZ8Hs6PPnBn738YyJG5QiDY8kLU_u5opnHMTOrX2eFGIc1lQpmhEoDb3N;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : true,
-          "subscriptionId" : "2c92c0fb525e37ea01526120f38f73db",
-          "subscriptionNumber" : "A-S00000044",
-          "contractedMrr" : 59.000000000,
-          "totalContractedValue" : null,
-          "invoiceId" : "2c92c0fb525e37ea01526120f47573e4",
-          "paymentId" : "2c92c0fb525e37ea01526120f4d873ef",
-          "paidAmount" : 59.00
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:21:35 GMT
-recorded_with: VCR 3.0.1
