--- spec/cassettes/update_subscription_quantity_failure.yml
@@ -1,110 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://apisandbox-api.zuora.com/rest/v1/connections
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Apiaccesskeyid:
-      - engineering+dev@contactually.com
-      Apisecretaccesskey:
-      - "<ZUORA_SANDBOX_PASSWORD>"
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '0'
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '22'
-      Expires:
-      - Wed, 20 Jan 2016 22:42:00 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:42:00 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=_2VjZwInSoQnx__tl3prC9ufgSkOMWBtLPh2gUqaA0geRSkKrXAKyrx2mIkB6Uc6KAifT5JnTe-1jmXiFJXJskubuoYrOs1N2Q00TwTh5VAgNC5m-Yj8WYNNUIT9bW9ZLwxtFHiTtrfjbcqDhrRlRP2pNf7p0NJqGHYDwdniIooGazfIKVMqfe5rlJLjshC2G3I6ggvWfvAiwN-LEgpZ2vI1zifGM-d4by_KvEJG32tGVCd1TPiGDTCkUc9worIn;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : true
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:41:59 GMT
-- request:
-    method: put
-    uri: https://apisandbox-api.zuora.com/rest/v1/subscriptions/2c92c0fa525e261a01526127f6416c8b
-    body:
-      encoding: UTF-8
-      string: '{"invoiceCollect":true,"applyCreditBalance":true,"update":[{"ratePlanId":"2c92c0fa525e261a01526127f6496c8f","contractEffectiveDate":"2016-01-20","chargeUpdateDetails":[{"ratePlanChargeId":"2c92c0fa525e261a01526127f6496c90","quantity":"2"}]}]}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Content-Type:
-      - application/json
-      Cookie:
-      - ZSession=_2VjZwInSoQnx__tl3prC9ufgSkOMWBtLPh2gUqaA0geRSkKrXAKyrx2mIkB6Uc6KAifT5JnTe-1jmXiFJXJskubuoYrOs1N2Q00TwTh5VAgNC5m-Yj8WYNNUIT9bW9ZLwxtFHiTtrfjbcqDhrRlRP2pNf7p0NJqGHYDwdniIooGazfIKVMqfe5rlJLjshC2G3I6ggvWfvAiwN-LEgpZ2vI1zifGM-d4by_KvEJG32tGVCd1TPiGDTCkUc9worIn;
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '170'
-      Expires:
-      - Wed, 20 Jan 2016 22:42:01 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:42:01 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=Us3oJG-4KyOJwc5bdi1dArgFLkrdSFmQ4DYAf4nR-rFkGAKzTA-37SeommpO9P-jaVuqCQ55j0kgJnjJU7IIOg5IF52V1aDFm5czWoZTZOlcCTm8H7OPdZxWdDBVzgjTfRsmVLqDdFgSztAdLMIfjJsUaCR82JKxzyBJZgZBMhriA9Xc1GFDWVnso-uZ6Zwo9_jfn0_U1pk2GydQw4jOrOVOwuLfl-puHoz-9nXoZopYprOXen45Kbs3Vx-CsiG3;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : false,
-          "processId" : "A0FA991803CC26CE",
-          "reasons" : [ {
-            "code" : 53500030,
-            "message" : "This subscription is not the latest version."
-          } ]
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:42:00 GMT
-recorded_with: VCR 3.0.1
