--- spec/cassettes/update_subscription_quantity_success.yml
@@ -1,111 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://apisandbox-api.zuora.com/rest/v1/connections
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Apiaccesskeyid:
-      - engineering+dev@contactually.com
-      Apisecretaccesskey:
-      - "<ZUORA_SANDBOX_PASSWORD>"
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '0'
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '22'
-      Expires:
-      - Wed, 20 Jan 2016 22:41:58 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:41:58 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=s3vZajwk-juAn8sR5fpQjAErz3wxIFwwzGrCgRbhVs844L6xs0rQdBtWSz6mV3KMVW3YyTgPZzQ6xRV4hACygVu3zr81f82hoZ00CuNIXKh2gHsL8L8OECZxz3SqZW96LVA_iVHjg9ky1FRGRmOgM0NC-a2IKKKTHb4ZPAJaPFKWiRrKBqq8YotM0kD5W4UfBmWKjHrcXBgsXg5R6UuvX9qDLa86j4128ZzGp-wu2su2eSeh0f1tO_JvozIMMku9;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : true
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:41:57 GMT
-- request:
-    method: put
-    uri: https://apisandbox-api.zuora.com/rest/v1/subscriptions/2c92c0fa525e261a01526127f6416c8b
-    body:
-      encoding: UTF-8
-      string: '{"invoiceCollect":true,"applyCreditBalance":true,"update":[{"ratePlanId":"2c92c0fa525e261a01526127f6496c8f","contractEffectiveDate":"2016-01-20","chargeUpdateDetails":[{"ratePlanChargeId":"2c92c0fa525e261a01526127f6496c90","quantity":"2"}]}]}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Content-Type:
-      - application/json
-      Cookie:
-      - ZSession=s3vZajwk-juAn8sR5fpQjAErz3wxIFwwzGrCgRbhVs844L6xs0rQdBtWSz6mV3KMVW3YyTgPZzQ6xRV4hACygVu3zr81f82hoZ00CuNIXKh2gHsL8L8OECZxz3SqZW96LVA_iVHjg9ky1FRGRmOgM0NC-a2IKKKTHb4ZPAJaPFKWiRrKBqq8YotM0kD5W4UfBmWKjHrcXBgsXg5R6UuvX9qDLa86j4128ZzGp-wu2su2eSeh0f1tO_JvozIMMku9;
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '267'
-      Expires:
-      - Wed, 20 Jan 2016 22:41:59 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:41:59 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=PJ4HEiKl4gSk-MPggp3TlG_vI13NpywKuPuS0G6t9K6GCAaF0FBlx-0VENacyS5r-yCabQqAdQkS3zC1_ONNEMUgQNhFxCdhyw1CUYp2_SLCZ4Pi9RxscAihQvSoEOQMrAqkvHRm4sWhO2EdfsfwYtt-my96FHP0N2cJvhv3xhy-si7d4Vsf7_Lv_jaViXr1QKvo2zBM3mHyEzBPz_W_2UQHFo6yvyXDaDKzn36z0Cy2srLCVLExcMgwZm2B43Ve;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : true,
-          "subscriptionId" : "2c92c0fa525e261a015261339d826dbe",
-          "totalDeltaMrr" : 35.000000000,
-          "totalDeltaTcv" : null,
-          "invoiceId" : "2c92c0fa525e261a015261339e246dce",
-          "paymentId" : "2c92c0fa525e261a015261339e976ddb",
-          "paidAmount" : 35.00
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:41:59 GMT
-recorded_with: VCR 3.0.1
