--- spec/cassettes/update_subscription_rate_plan_success.yml
@@ -1,109 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://apisandbox-api.zuora.com/rest/v1/connections
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Apiaccesskeyid:
-      - engineering+dev@contactually.com
-      Apisecretaccesskey:
-      - "<ZUORA_SANDBOX_PASSWORD>"
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '0'
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '22'
-      Expires:
-      - Wed, 20 Jan 2016 22:29:14 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:29:14 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=7DCNFnQ_OKD0Xswua6XG0_31FcEMY5cgGkoeXSymaxov0kCcHliMRp_RZbxsmraEy_skmsrznXxWVm8F5Mi3kbrkE-mkWmnU1TmRkcAa7bJaOgbZoJt5H7hZoGuiQRygW_2VSahP_VhXmu4RA6nMEmRFbaLwvPrSfsZO0GRku6WBbpYkf-l0hMGM-7n81X7wo_v0rrN5KhHDiVbroR6wfxtHYvd7033_sRDlY2CTJcY3XKhS1WJXCXmhxOaFkuGk;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : true
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:29:13 GMT
-- request:
-    method: put
-    uri: https://apisandbox-api.zuora.com/rest/v1/subscriptions/2c92c0fb525e37ea01526120f38f73db
-    body:
-      encoding: UTF-8
-      string: '{"invoiceCollect":true,"applyCreditBalance":true,"remove":[{"ratePlanId":"2c92c0fb525e37ea01526120f37f73d9","contractEffectiveDate":"2016-01-20"}],"add":[{"productRatePlanId":"2c92c0f850fa443e01510cbb557e3cf8","contractEffectiveDate":"2016-01-20"}]}'
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Content-Type:
-      - application/json
-      Cookie:
-      - ZSession=7DCNFnQ_OKD0Xswua6XG0_31FcEMY5cgGkoeXSymaxov0kCcHliMRp_RZbxsmraEy_skmsrznXxWVm8F5Mi3kbrkE-mkWmnU1TmRkcAa7bJaOgbZoJt5H7hZoGuiQRygW_2VSahP_VhXmu4RA6nMEmRFbaLwvPrSfsZO0GRku6WBbpYkf-l0hMGM-7n81X7wo_v0rrN5KhHDiVbroR6wfxtHYvd7033_sRDlY2CTJcY3XKhS1WJXCXmhxOaFkuGk;
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - Zuora App
-      Content-Type:
-      - application/json;charset=utf-8
-      Content-Length:
-      - '192'
-      Expires:
-      - Wed, 20 Jan 2016 22:29:15 GMT
-      Cache-Control:
-      - max-age=0, no-cache, no-store
-      Pragma:
-      - no-cache
-      Date:
-      - Wed, 20 Jan 2016 22:29:15 GMT
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - ZSession=Je6TEVZ5S1G8QHLCqenDf89Il1LmpRi6KUOcc63p7BLT518xOpgJW7dLglznOxrjeGPRVTfYcPw5rpq5qypS-x417U5ZbHg-193N1sY45qbu6htYPFAnn1litArXKQHDPGak9dbbAs9V0Z1_i3Ohx1nALZ1OC7A_rEH1kV3Pm7IhV63GlU1D_x3P-CwBJtwStLoHf2uPpZO1dF5NNd2czVxg-eVCHZF2991PTPUBSJgxen1RB4nKgxyBIkSDiCEJ;
-        Path=/; Secure; HttpOnly
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "success" : true,
-          "subscriptionId" : "2c92c0fa525e261a01526127f6416c8b",
-          "totalDeltaMrr" : -24.000000000,
-          "totalDeltaTcv" : null,
-          "invoiceId" : "2c92c0fa525e261a01526127f6cb6c98"
-        }
-    http_version: 
-  recorded_at: Wed, 20 Jan 2016 22:29:14 GMT
-recorded_with: VCR 3.0.1
