+++ spec/integration/connection_spec.rb
@@ -1,13 +1,13 @@
 require 'spec_helper'
 require 'nokogiri'
 
-describe Zuora::Soap::Client do
+describe Zuora::Client do
   let(:username) { ENV['ZUORA_SANDBOX_USERNAME'] }
   let(:password) { ENV['ZUORA_SANDBOX_PASSWORD'] }
   let(:vcr_options) do
     { match_requests_on: [:path] }
   end
-  let(:client) { Zuora::Soap::Client.new(username, password, true) }
+  let(:client) { Zuora::Client.new(username, password, true) }
 
   let(:soap_success_xpath) do
     '/soapenv:Envelope/soapenv:Body/ns1:createResponse/ns1:result/ns1:Success'
@@ -49,7 +49,7 @@ describe Zuora::Soap::Client do
 
     let(:create_bill_run_success_status) do
       Nokogiri::XML(create_bill_run_success_response.body).xpath(
-        soap_success_xpath, Zuora::Soap::NAMESPACES
+        soap_success_xpath, Zuora::NAMESPACES
       ).text
     end
 
@@ -65,7 +65,7 @@ describe Zuora::Soap::Client do
 
     let(:create_bill_run_failure_status) do
       Nokogiri::XML(create_bill_run_failure_response.body).xpath(
-        soap_success_xpath, Zuora::Soap::NAMESPACES
+        soap_success_xpath, Zuora::NAMESPACES
       ).text
     end
 
@@ -96,7 +96,7 @@ describe Zuora::Soap::Client do
     let(:parse_success) do
       lambda do |xml|
         Nokogiri::XML(xml).xpath(
-          soap_success_xpath, Zuora::Soap::NAMESPACES
+          soap_success_xpath, Zuora::NAMESPACES
         ).text
       end
     end
@@ -119,7 +119,7 @@ describe Zuora::Soap::Client do
 
     let(:create_refund_failure_status) do
       Nokogiri::XML(create_refund_failure_response.body).xpath(
-        soap_success_xpath, Zuora::Soap::NAMESPACES
+        soap_success_xpath, Zuora::NAMESPACES
       ).text
     end
