+++ spec/spec_helper.rb
@@ -5,7 +5,6 @@ if ENV['CIRCLECI']
 end
 
 require_relative '../lib/zuora'
-require_relative '../lib/utils/schema_model'
 
 require 'byebug'
 require 'factory_girl'
